Synthetic Data Generation/Singan-Seg/requirements.txt
# pip install -r requirements.txt

imageio==2.36.0
lazy_loader==0.4
networkx==3.4.2
numpy==2.1.3
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
pillow==11.0.0
python-dateutil==2.9.0.post0
pytz==2024.2
scikit-image==0.24.0
scipy==1.14.1
six==1.16.0
tifffile==2024.9.20
tzdata==2024.2
torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
matplotlib
scikit-learn